requirements.txt
google-adk==0.3.0
google-generativeai==0.8.5
python-dotenv==1.1.0
streamlit==1.36.0
scikit-learn==1.5.0
numpy>=1.26.0
pandas>=2.2.0
rich>=13.7.0
deprecated>=1.2.14 